requirements.txt
pillow>=10.0.0
numpy
torch>=2.0.0